submission

+++ heap.c
@@ -210,8 +210,8 @@ void lhu_h(blob* p_vm, uint32_t reg_index, uint32_t address) {
 }
 
 
-uint8_t get_num_banks(uint32_t malloc_size) {
+uint32_t get_num_banks(uint32_t malloc_size) {
     if (malloc_size % 64 == 0)
-        return (uint8_t) malloc_size / 64;
-    return (uint8_t) (malloc_size / 64) + 1;
+        return (uint32_t) malloc_size / 64;
+    return (uint32_t) (malloc_size / 64) + 1;
 }
